requirements.txt
requests
pyyaml
python-dotenv
docker
python-slugify
